More splits ; copyright ; cleanup
