+++ tests/test_client_library.py
@@ -17,7 +17,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
-"""Tests for ClientLibrary initialization, authentication, versioning, and lab management."""
+"""Tests for ClientLibrary authentication, lab management, and diagnostics."""
 
 from __future__ import annotations
 
@@ -25,7 +25,7 @@ import json
 import logging
 from collections.abc import Iterator
 from pathlib import Path
-from unittest.mock import MagicMock, call, patch
+from unittest.mock import MagicMock, patch
 
 import httpx
 import pytest
@@ -34,7 +34,6 @@ import respx
 from virl2_client.exceptions import APIError, LabNotFound
 from virl2_client.models import Lab
 from virl2_client.virl2_client import (
-    ClientConfig,
     ClientLibrary,
     DiagnosticsCategory,
     InitializationError,
@@ -43,26 +42,6 @@ from virl2_client.virl2_client import (
 
 CURRENT_VERSION = ClientLibrary.VERSION.version_str
 FAKE_URL = "https://0.0.0.0/fake_url/"
-
-
-@pytest.fixture
-def reset_env(monkeypatch: pytest.MonkeyPatch) -> None:
-    """Clear VIRL2-related environment variables for isolated init tests.
-
-    :param monkeypatch: Pytest monkeypatch fixture.
-    """
-    env_vars = [
-        "VIRL2_URL",
-        "VIRL_HOST",
-        "VIRL2_USER",
-        "VIRL_USERNAME",
-        "VIRL2_PASS",
-        "VIRL_PASSWORD",
-        "VIRL2_JWT",
-    ]
-
-    for key in env_vars:
-        monkeypatch.delenv(key, raising=False)
 
 
 @pytest.mark.parametrize("title", [None, "new_title"], ids=["default", "custom_title"])
@@ -96,47 +75,6 @@ def test_import_lab_from_path_virl(
         params=expected_params,
         content="<?xml version='1.0' encoding='UTF-8'?>",
     )
-
-
-def test_ssl_certificate(
-    client_library_server_current: MagicMock, mocked_session: MagicMock
-) -> None:
-    """Use constructor-provided SSL CA bundle path for requests.
-
-    :param client_library_server_current: Patched current-version fixture.
-    :param mocked_session: Mocked HTTP session fixture.
-    """
-    _ = client_library_server_current, mocked_session
-    cl = ClientLibrary(
-        url=FAKE_URL,
-        username="test",
-        password="pa$$",
-        ssl_verify="/home/user/cert.pem",
-    )
-    cl.is_system_ready(wait=True)
-
-    assert cl._ssl_verify == "/home/user/cert.pem"
-    assert cl._session.mock_calls[0] == call.get("authentication")
-
-
-def test_ssl_certificate_from_env_variable(
-    client_library_server_current: MagicMock,
-    monkeypatch: pytest.MonkeyPatch,
-    mocked_session: MagicMock,
-) -> None:
-    """Use CA_BUNDLE environment variable for SSL verification.
-
-    :param client_library_server_current: Patched current-version fixture.
-    :param monkeypatch: Fixture for temporary environment mutation.
-    :param mocked_session: Mocked HTTP session fixture.
-    """
-    _ = client_library_server_current, mocked_session
-    monkeypatch.setenv("CA_BUNDLE", "/home/user/cert.pem")
-    cl = ClientLibrary(url=FAKE_URL, username="test", password="pa$$")
-
-    assert cl.is_system_ready()
-    assert cl._ssl_verify == "/home/user/cert.pem"
-    assert cl._session.mock_calls[0] == call.get("authentication")
 
 
 @respx.mock
@@ -375,41 +313,6 @@ def test_old_auth_url_used_with_cml_2_9(
     assert not new_auth_route.called
 
 
-def test_client_library_init_allow_http(
-    client_library_server_current: MagicMock,
-) -> None:
-    """Client accepts http:// URL when allow_http=True.
-
-    :param client_library_server_current: Patched system_info fixture.
-    """
-    _ = client_library_server_current
-    cl = ClientLibrary("http://somehost", "virl2", "virl2", allow_http=True)
-    assert cl._session.base_url.scheme == "http"
-    assert cl._session.base_url.host == "somehost"
-    assert cl._session.base_url.port is None
-    assert cl._session.base_url.path.endswith("/api/v0/")
-    assert cl.username == "virl2"
-    assert cl.password == "virl2"
-
-
-@pytest.mark.parametrize("allow_http", [None, False], ids=["default", "explicit_false"])
-def test_init_disallow_http(
-    client_library_server_current: MagicMock,
-    allow_http: bool | None,
-) -> None:
-    """Client raises InitializationError for http:// when allow_http disallows.
-
-    NOTE: LLM-generated test -- verify for correctness.
-
-    :param client_library_server_current: Patched system_info fixture.
-    :param allow_http: Value for allow_http (None = omit kwarg).
-    """
-    _ = client_library_server_current
-    kwargs = {} if allow_http is None else {"allow_http": allow_http}
-    with pytest.raises(InitializationError, match="must be https"):
-        ClientLibrary("http://somehost", "virl2", "virl2", **kwargs)
-
-
 @respx.mock
 def test_new_auth_url_fails_with_cml_2_9(
     client_library_server_2_9_0: MagicMock,
@@ -473,181 +376,6 @@ def test_old_auth_url_deprecated_with_cml_2_10(
     assert not old_auth_route.called
 
 
-# the test fails if you have variables set in env
-@pytest.mark.parametrize("via", ["environment", "parameter"])
-@pytest.mark.parametrize("env_var", ["VIRL2_URL", "VIRL_HOST"])
-@pytest.mark.parametrize(
-    "params",
-    [
-        (False, "somehost"),
-        (False, "http://somehost"),
-        (False, "https://somehost:443"),
-        (True, "xyz://somehost:443"),
-        (True, "https:@somehost:4:4:3"),
-        (True, ""),
-        (True, None),
-    ],
-)
-def test_client_library_init_url(
-    client_library_server_current: MagicMock,
-    reset_env: None,
-    monkeypatch: pytest.MonkeyPatch,
-    via: str,
-    env_var: str,
-    params: tuple[bool, str | None],
-) -> None:
-    """ClientLibrary URL init from env or parameter with validation.
-
-    :param client_library_server_current: Patched system_info fixture.
-    :param reset_env: Fixture clearing VIRL2 env vars.
-    :param monkeypatch: Pytest monkeypatch fixture.
-    :param via: Source of URL ('environment' or 'parameter').
-    :param env_var: Environment variable name for URL.
-    :param params: Tuple of (should_fail, url_value).
-    """
-    _ = client_library_server_current, reset_env
-    monkeypatch.setattr("getpass.getpass", input)
-    (fail, url) = params
-    expected_parts = None if fail else httpx.URL(url)
-    if via == "environment":
-        env = url
-        url = None
-    else:
-        env = "http://badhost" if url else None
-    if env is None:
-        monkeypatch.delenv(env_var, raising=False)
-    else:
-        monkeypatch.setenv(env_var, env)
-    if fail:
-        with pytest.raises((InitializationError, OSError)) as err:
-            ClientLibrary(
-                url=url,
-                username="virl2",
-                password="virl2",
-                allow_http=True,
-                raise_for_auth_failure=True,
-            )
-        if isinstance(err.value, OSError):
-            assert "reading from stdin" in str(err.value)
-    else:
-        cl = ClientLibrary(url, username="virl2", password="virl2", allow_http=True)
-        url_parts = cl._session.base_url
-        assert url_parts.scheme == (expected_parts.scheme or "https")
-        assert url_parts.host == (expected_parts.host or expected_parts.path)
-        assert url_parts.port == expected_parts.port
-        assert url_parts.path == "/api/v0/"
-        assert cl._session.base_url.path.endswith("/api/v0/")
-        assert cl.username == "virl2"
-        assert cl.password == "virl2"
-
-
-# the test fails if you have variables set in env
-@pytest.mark.parametrize("via", ["environment", "parameter"])
-@pytest.mark.parametrize("env_var", ["VIRL2_USER", "VIRL_USERNAME"])
-@pytest.mark.parametrize("params", [(False, "johndoe"), (True, ""), (True, None)])
-def test_client_library_init_user(
-    client_library_server_current: MagicMock,
-    reset_env: None,
-    monkeypatch: pytest.MonkeyPatch,
-    via: str,
-    env_var: str,
-    params: tuple[bool, str | None],
-) -> None:
-    """ClientLibrary username init from env or parameter with validation.
-
-    :param client_library_server_current: Patched system_info fixture.
-    :param reset_env: Fixture clearing VIRL2 env vars.
-    :param monkeypatch: Pytest monkeypatch fixture.
-    :param via: Source of username ('environment' or 'parameter').
-    :param env_var: Environment variable name for username.
-    :param params: Tuple of (should_fail, username_value).
-    """
-    _ = client_library_server_current, reset_env
-    monkeypatch.setattr("getpass.getpass", input)
-    url = "validhostname"
-    (fail, user) = params
-    if via == "environment":
-        # can't set a None value for an environment variable
-        env = user or ""
-        user = None
-    else:
-        env = "baduser" if user else ""
-    if env is None:
-        monkeypatch.delenv(env_var, raising=False)
-    else:
-        monkeypatch.setenv(env_var, env)
-    if fail:
-        with pytest.raises((OSError, InitializationError)) as err:
-            ClientLibrary(url=url, username=user, password="virl2")
-        if isinstance(err.value, OSError):
-            assert "reading from stdin" in str(err.value)
-    else:
-        cl = ClientLibrary(url, username=user, password="virl2")
-        assert cl.username == params[1]
-        assert cl.password == "virl2"
-        assert cl._session.base_url == "https://validhostname/api/v0/"
-
-
-# the test fails if you have variables set in env
-@pytest.mark.parametrize("via", ["environment", "parameter"])
-@pytest.mark.parametrize("env_var", ["VIRL2_PASS", "VIRL_PASSWORD"])
-@pytest.mark.parametrize("params", [(False, "validPa$$w!2"), (True, ""), (True, None)])
-def test_client_library_init_password(
-    client_library_server_current: MagicMock,
-    reset_env: None,
-    monkeypatch: pytest.MonkeyPatch,
-    via: str,
-    env_var: str,
-    params: tuple[bool, str | None],
-) -> None:
-    """ClientLibrary password init from env or parameter with validation.
-
-    :param client_library_server_current: Patched system_info fixture.
-    :param reset_env: Fixture clearing VIRL2 env vars.
-    :param monkeypatch: Pytest monkeypatch fixture.
-    :param via: Source of password ('environment' or 'parameter').
-    :param env_var: Environment variable name for password.
-    :param params: Tuple of (should_fail, password_value).
-    """
-    _ = client_library_server_current, reset_env
-    monkeypatch.setattr("getpass.getpass", input)
-    url = "validhostname"
-    (fail, password) = params
-    if via == "environment":
-        # can't set a None value for an environment variable
-        env = password or ""
-        password = None
-    else:
-        env = "badpass" if password else ""
-    if env is None:
-        monkeypatch.delenv(env_var, raising=False)
-    else:
-        monkeypatch.setenv(env_var, env)
-    if fail:
-        with pytest.raises((OSError, InitializationError)) as err:
-            ClientLibrary(url=url, username="virl2", password=password)
-        if isinstance(err.value, OSError):
-            assert "reading from stdin" in str(err.value)
-    else:
-        cl = ClientLibrary(url, username="virl2", password=password)
-        assert cl.username == "virl2"
-        assert cl.password == params[1]
-        assert cl._session.base_url == "https://validhostname/api/v0/"
-
-
-def test_client_library_str_and_repr(
-    client_library_server_current: MagicMock,
-) -> None:
-    """ClientLibrary str and repr return expected format.
-
-    :param client_library_server_current: Patched system_info fixture.
-    """
-    _ = client_library_server_current
-    client_library = ClientLibrary("somehost", "virl2", password="virl2")
-    assert repr(client_library) == "ClientLibrary('https://somehost')"
-    assert str(client_library) == "ClientLibrary URL: https://somehost/api/v0/"
-
-
 def test_incompatible_version(
     client_library_server_2_0_0: MagicMock,
 ) -> None:
@@ -695,374 +423,6 @@ def test_client_minor_version_lt_warn(
         f"Please ensure the client version is compatible with the controller version. "
         f"Client {CURRENT_VERSION}, controller 2.19.0." in caplog.text
     )
-
-
-@pytest.mark.parametrize(
-    "a, b, expected",
-    [
-        pytest.param(Version("2.0.0"), Version("2.0.0"), True, id="equal"),
-        pytest.param(Version("2.0.0"), Version("2.0.1"), False, id="differ"),
-        pytest.param(Version("2.0.0"), "2.0.0", False, id="string"),
-        pytest.param(Version("2.0.0"), 200, False, id="int"),
-    ],
-)
-def test_version_comparison_eq(
-    a: Version, b: Version | str | int, expected: bool
-) -> None:
-    """Compare Version objects with equality operator.
-
-    NOTE: LLM-generated test -- verify for correctness.
-
-    :param a: First operand.
-    :param b: Second operand.
-    :param expected: Expected result of a == b.
-    """
-    assert (a == b) == expected
-
-
-@pytest.mark.parametrize(
-    "greater, lesser, expected",
-    [
-        pytest.param(
-            Version("2.0.1"), Version("2.0.0"), True, id="Patch is greater than"
-        ),
-        pytest.param(
-            Version("2.0.10"), Version("2.0.0"), True, id="Patch is much greater than"
-        ),
-        pytest.param(
-            Version("2.1.0"), Version("2.0.0"), True, id="Minor is greater than"
-        ),
-        pytest.param(
-            Version("2.10.0"), Version("2.0.0"), True, id="Minor is much greater than"
-        ),
-        pytest.param(
-            Version("3.0.0"), Version("2.0.0"), True, id="Major is greater than"
-        ),
-        pytest.param(
-            Version("10.0.0"), Version("2.0.0"), True, id="Major is much greater than"
-        ),
-        pytest.param(
-            Version("2.0.0"), Version("2.0.1"), False, id="Patch is lesser than"
-        ),
-        pytest.param(
-            Version("2.0.0"), Version("2.0.10"), False, id="Patch is much lesser than"
-        ),
-        pytest.param(
-            Version("2.0.0"), Version("2.1.0"), False, id="Minor is lesser than"
-        ),
-        pytest.param(
-            Version("2.0.0"), Version("2.10.0"), False, id="Minor is much lesser than"
-        ),
-        pytest.param(
-            Version("2.0.0"), Version("3.0.0"), False, id="Major is lesser than"
-        ),
-        pytest.param(
-            Version("2.0.0"), Version("10.0.0"), False, id="Major is much lesser than"
-        ),
-        pytest.param(
-            Version("2.0.0"),
-            "random string",
-            False,
-            id="Other object is string and not a Version object",
-        ),
-        pytest.param(
-            Version("2.0.0"),
-            12345,
-            False,
-            id="Other object is int and not a Version object",
-        ),
-    ],
-)
-def test_version_comparison_gt(
-    greater: Version, lesser: Version | str | int, expected: bool
-) -> None:
-    """Compare Version objects with greater-than operator.
-
-    NOTE: LLM-generated test -- verify for correctness.
-
-    :param greater: Version expected to be greater.
-    :param lesser: Version or other object to compare against.
-    :param expected: Expected result of greater > lesser.
-    """
-    assert (greater > lesser) == expected
-
-
-@pytest.mark.parametrize(
-    "first, second, expected",
-    [
-        pytest.param(
-            Version("2.0.1"), Version("2.0.0"), True, id="Patch is greater than"
-        ),
-        pytest.param(
-            Version("2.0.10"), Version("2.0.0"), True, id="Patch is much greater than"
-        ),
-        pytest.param(
-            Version("2.1.0"), Version("2.0.0"), True, id="Minor is greater than"
-        ),
-        pytest.param(
-            Version("2.10.0"), Version("2.0.0"), True, id="Minor is much greater than"
-        ),
-        pytest.param(
-            Version("3.0.0"), Version("2.0.0"), True, id="Major is greater than"
-        ),
-        pytest.param(
-            Version("10.0.0"), Version("2.0.0"), True, id="Major is much greater than"
-        ),
-        pytest.param(
-            Version("2.0.0"), Version("2.0.1"), False, id="Patch is lesser than"
-        ),
-        pytest.param(
-            Version("2.0.0"), Version("2.0.10"), False, id="Patch is much lesser than"
-        ),
-        pytest.param(
-            Version("2.0.0"), Version("2.1.0"), False, id="Minor is lesser than"
-        ),
-        pytest.param(
-            Version("2.0.0"), Version("2.10.0"), False, id="Minor is much lesser than"
-        ),
-        pytest.param(
-            Version("2.0.0"), Version("3.0.0"), False, id="Major is lesser than"
-        ),
-        pytest.param(
-            Version("2.0.0"), Version("10.0.0"), False, id="Major is much lesser than"
-        ),
-        pytest.param(
-            Version("2.0.0"),
-            Version("2.0.0"),
-            True,
-            id="Equal versions no minor no patch",
-        ),
-        pytest.param(
-            Version("2.0.1"),
-            Version("2.0.1"),
-            True,
-            id="Equal versions patch increment",
-        ),
-        pytest.param(
-            Version("2.1.0"),
-            Version("2.1.0"),
-            True,
-            id="Equal versions minor increment",
-        ),
-        pytest.param(
-            Version("3.0.0"),
-            Version("3.0.0"),
-            True,
-            id="Equal versions major increment",
-        ),
-        pytest.param(
-            Version("2.0.0"),
-            "random string",
-            False,
-            id="Other object is string and not a Version object",
-        ),
-        pytest.param(
-            Version("2.0.0"),
-            12345,
-            False,
-            id="Other object is int and not a Version object",
-        ),
-    ],
-)
-def test_version_comparison_gte(
-    first: Version, second: Version, expected: bool
-) -> None:
-    """Compare Version objects with greater-than-or-equal operator.
-
-    NOTE: LLM-generated test -- verify for correctness.
-
-    :param first: First Version to compare.
-    :param second: Second Version to compare against.
-    :param expected: Expected result of first >= second.
-    """
-    assert (first >= second) == expected
-
-
-@pytest.mark.parametrize(
-    "lesser, greater, expected",
-    [
-        pytest.param(Version("2.0.0"), Version("2.0.1"), True, id="Patch is less than"),
-        pytest.param(
-            Version("2.0.0"), Version("2.0.10"), True, id="Patch is much less than"
-        ),
-        pytest.param(Version("2.0.0"), Version("2.1.0"), True, id="Minor is less than"),
-        pytest.param(
-            Version("2.0.0"), Version("2.10.0"), True, id="Minor is much less than"
-        ),
-        pytest.param(Version("2.0.0"), Version("3.0.0"), True, id="Major is less than"),
-        pytest.param(
-            Version("2.0.0"), Version("10.0.0"), True, id="Major is much less than"
-        ),
-        pytest.param(
-            Version("2.0.1"), Version("2.0.0"), False, id="Patch is greater than"
-        ),
-        pytest.param(
-            Version("2.0.10"), Version("2.0.0"), False, id="Patch is much greater than"
-        ),
-        pytest.param(
-            Version("2.1.0"), Version("2.0.0"), False, id="Minor is greater than"
-        ),
-        pytest.param(
-            Version("2.10.0"), Version("2.0.0"), False, id="Minor is much greater than"
-        ),
-        pytest.param(
-            Version("3.0.0"), Version("2.0.0"), False, id="Major is greater than"
-        ),
-        pytest.param(
-            Version("10.0.0"), Version("2.0.0"), False, id="Major is much greater than"
-        ),
-        pytest.param(
-            Version("2.0.0"),
-            "random string",
-            False,
-            id="Other object is string and not a Version object",
-        ),
-        pytest.param(
-            Version("2.0.0"),
-            12345,
-            False,
-            id="Other object is int and not a Version object",
-        ),
-    ],
-)
-def test_version_comparison_lt(
-    lesser: Version, greater: Version | str | int, expected: bool
-) -> None:
-    """Compare Version objects with less-than operator.
-
-    NOTE: LLM-generated test -- verify for correctness.
-
-    :param lesser: Version expected to be lesser.
-    :param greater: Version or other object to compare against.
-    :param expected: Expected result of lesser < greater.
-    """
-    assert (lesser < greater) == expected
-
-
-@pytest.mark.parametrize(
-    "first, second, expected",
-    [
-        pytest.param(Version("2.0.0"), Version("2.0.1"), True, id="Patch is less than"),
-        pytest.param(
-            Version("2.0.0"), Version("2.0.10"), True, id="Patch is much less than"
-        ),
-        pytest.param(Version("2.0.0"), Version("2.1.0"), True, id="Minor is less than"),
-        pytest.param(
-            Version("2.0.0"), Version("2.10.0"), True, id="Minor is much less than"
-        ),
-        pytest.param(Version("2.0.0"), Version("3.0.0"), True, id="Major is less than"),
-        pytest.param(
-            Version("2.0.0"), Version("10.0.0"), True, id="Major is much less than"
-        ),
-        pytest.param(
-            Version("2.0.1"), Version("2.0.0"), False, id="Patch is greater than"
-        ),
-        pytest.param(
-            Version("2.0.10"), Version("2.0.0"), False, id="Patch is much greater than"
-        ),
-        pytest.param(
-            Version("2.1.0"), Version("2.0.0"), False, id="Minor is greater than"
-        ),
-        pytest.param(
-            Version("2.10.0"), Version("2.0.0"), False, id="Minor is much greater than"
-        ),
-        pytest.param(
-            Version("3.0.0"), Version("2.0.0"), False, id="Major is greater than"
-        ),
-        pytest.param(
-            Version("10.0.0"), Version("2.0.0"), False, id="Major is much greater than"
-        ),
-        pytest.param(
-            Version("2.0.0"),
-            Version("2.0.0"),
-            True,
-            id="Equal versions no minor no patch",
-        ),
-        pytest.param(
-            Version("2.0.1"),
-            Version("2.0.1"),
-            True,
-            id="Equal versions patch increment",
-        ),
-        pytest.param(
-            Version("2.1.0"),
-            Version("2.1.0"),
-            True,
-            id="Equal versions minor increment",
-        ),
-        pytest.param(
-            Version("3.0.0"),
-            Version("3.0.0"),
-            True,
-            id="Equal versions major increment",
-        ),
-        pytest.param(
-            Version("2.0.0"),
-            "random string",
-            False,
-            id="Other object is string and not a Version object",
-        ),
-        pytest.param(
-            Version("2.0.0"),
-            12345,
-            False,
-            id="Other object is int and not a Version object",
-        ),
-    ],
-)
-def test_version_comparison_lte(
-    first: Version, second: Version, expected: bool
-) -> None:
-    """Compare Version objects with less-than-or-equal operator.
-
-    NOTE: LLM-generated test -- verify for correctness.
-
-    :param first: First Version to compare.
-    :param second: Second Version to compare against.
-    :param expected: Expected result of first <= second.
-    """
-    assert (first <= second) == expected
-
-
-@pytest.mark.parametrize(
-    "version_str",
-    [
-        "2.1.0-dev0+build8.7ee86bf8",
-        "2.1.0dev0+build8.7ee86bf8",
-        "2.1.0--dev0+build8.7ee86bf8",
-        "2.1.0_dev0+build8.7ee86bf8",
-        "2.1.0",
-        "2.1.0-",
-    ],
-)
-def test_version_parse_valid(version_str: str) -> None:
-    """Parse valid Version string formats into expected components.
-
-    NOTE: LLM-generated test -- verify for correctness.
-
-    :param version_str: Version string to parse.
-    """
-    v = Version(version_str)
-    assert v.major == 2 and v.minor == 1 and v.patch == 0
-
-
-@pytest.mark.parametrize(
-    "version_str",
-    [
-        "2.1-dev0+build8.7ee86bf8",
-        "2-dev0+build8.7ee86bf8",
-        "54dev0+build8.7ee86bf8",
-    ],
-)
-def test_version_parse_invalid(version_str: str) -> None:
-    """Reject malformed version strings with ValueError.
-
-    NOTE: LLM-generated test -- verify for correctness.
-
-    :param version_str: Invalid version string.
-    """
-    with pytest.raises(ValueError):
-        Version(version_str)
 
 
 def test_import_lab_rejects_offline_argument(
@@ -1393,69 +753,6 @@ def test_join_existing_lab_404_not_found(
 
     with pytest.raises(LabNotFound):
         cl.join_existing_lab(lab_id)
-
-
-def test_version_diff_helpers() -> None:
-    """Version helper methods return True when versions differ.
-
-    NOTE: LLM-generated test -- verify for correctness.
-    """
-    v1 = Version("2.10.3")
-    v2 = Version("3.11.4")
-    assert v1.major_differs(v2) is True
-    assert v1.minor_differs(v2) is True
-    assert v1.patch_differs(v2) is True
-    assert v1.minor_or_patch_differs(v2) is True
-
-
-def test_config_get_from_file(tmp_path: Path) -> None:
-    """ClientConfig._get_from_file reads property from .virlrc.
-
-    NOTE: LLM-generated test -- verify for correctness.
-
-    :param tmp_path: Temporary directory fixture.
-    """
-    config_file = tmp_path / ".virlrc"
-    config_file.write_text('VIRL2_URL="https://from-file"\n')
-    assert ClientConfig._get_from_file(tmp_path, "VIRL2_URL") == "https://from-file"
-
-
-def test_config_get_prop(tmp_path: Path, monkeypatch: pytest.MonkeyPatch) -> None:
-    """ClientConfig._get_prop walks directory tree to find .virlrc.
-
-    NOTE: LLM-generated test -- verify for correctness.
-
-    :param tmp_path: Temporary directory fixture.
-    :param monkeypatch: Pytest monkeypatch fixture.
-    """
-    config_file = tmp_path / ".virlrc"
-    config_file.write_text('VIRL2_URL="https://from-file"\n')
-    nested = tmp_path / "a" / "b"
-    nested.mkdir(parents=True)
-    monkeypatch.chdir(nested)
-    assert ClientConfig._get_prop("VIRL2_URL") == "https://from-file"
-
-
-def test_config_populate_inputs(monkeypatch: pytest.MonkeyPatch) -> None:
-    """ClientConfig._populate_from_inputs stores JWT from interactive input.
-
-    NOTE: LLM-generated test -- verify for correctness.
-
-    :param monkeypatch: Pytest monkeypatch fixture.
-    """
-    conf = {
-        "url": None,
-        "username": None,
-        "password": None,
-        "jwtoken": None,
-        "ssl_verify": True,
-    }
-    monkeypatch.setattr(
-        "builtins.input",
-        MagicMock(side_effect=["https://server.local", "x" * 40]),
-    )
-    ClientConfig._populate_from_inputs(conf)
-    assert conf["jwtoken"] == "x" * 40
 
 
 def test_client_uuid(
